+++ contracts/asset/JNT/JNTController.sol
@@ -29,7 +29,7 @@ contract JNTController is CrydrControllerBase,
 
   /* JNTControllerInterface */
 
-  function chargeJNT(address _from, address _to, uint _value) onlyAllowedManager('jnt_payable_service') {
+  function chargeJNT(address _from, address _to, uint256 _value) onlyAllowedManager('jnt_payable_service') {
     CrydrStorageERC20Interface(address(crydrStorage)).transfer(_from, _to, _value);
     JNTChargedEvent(msg.sender, _from, _to, _value);
   }
