+++ contracts/crydr/controller/CrydrControllerMintableInterface.sol
@@ -13,6 +13,6 @@ contract CrydrControllerMintableInterface {
 
   // todo views should know about Mint/Burn events
 
-  function mint(address _account, uint _value);
-  function burn(address _account, uint _value);
+  function mint(address _account, uint256 _value);
+  function burn(address _account, uint256 _value);
 }
