+++ contracts/crydr/storage/CrydrStorageERC20Interface.sol
@@ -11,14 +11,14 @@ contract CrydrStorageERC20Interface {
 
   /* Events */
 
-  event CrydrTransferredEvent(address indexed from, address indexed to, uint value);
-  event CrydrTransferredFromEvent(address indexed spender, address indexed from, address indexed to, uint value);
-  event CrydrSpendingApprovedEvent(address indexed owner, address indexed spender, uint value);
+  event CrydrTransferredEvent(address indexed from, address indexed to, uint256 value);
+  event CrydrTransferredFromEvent(address indexed spender, address indexed from, address indexed to, uint256 value);
+  event CrydrSpendingApprovedEvent(address indexed owner, address indexed spender, uint256 value);
 
 
   /* ERC20 optimization. _msgsender - account that invoked CrydrView */
 
-  function transfer(address _msgsender, address _to, uint _value);
-  function transferFrom(address _msgsender, address _from, address _to, uint _value);
-  function approve(address _msgsender, address _spender, uint _value);
+  function transfer(address _msgsender, address _to, uint256 _value);
+  function transferFrom(address _msgsender, address _from, address _to, uint256 _value);
+  function approve(address _msgsender, address _spender, uint256 _value);
 }
